DO_NOT_USE_EVER
This commit adds a mechanism to retrieve logcat and dmesg when the
device is not booting and adb is not working/not yet enabled during the
boot process.

It is bad but this is how it works:
- we have a service in init (start_dmesg)
- this service can start when you want -> in this commit, I start it at
  post-fs-data and immediately before the hwservicemanager. Keep in mind
that the service does not even exit (see next point).
- the service copies the dmesg output and the logcat output to two files
  (dmesg.txt and logcat.txt) in the data partition - it also mounts the
data partitions, so it can work before data is mounted by init. The
service copies the dmesg every seconds, overwriting the previous file:
while doing this, it creates dmesg1.txt to avoid truncated files in case
of a reboot during the copy - the logcat does not need this, as it is
continuously updated.
- it also copy the output of the getprop command in the getprop.txt
  file.

As you can image, it does not work with SELinux in enforcing and I
didn't care to create a label for it. It should be used only for
debugging issues causing the phone to reboot very early during the boot
process.

It is bad and you should not use.

If you are reading this last line, you are very desperate my friend.
Good luck debugging your issue, I'm with you.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>